Для каждого компонента создал отдельную папку и выделил класс и шаблон в 2 разных файла
