+++ src/components/board/board-tpl.js
@@ -1,0 +1,13 @@
+/**
+ * Создаем разметку блока Доски
+ * @return {string} Разметка
+ */
+const createBoardTemplate = () => {
+  return (
+    `<section class="board container"></section>`
+  );
+};
+
+export {
+  createBoardTemplate
+};
