+++ src/components/filter/filter.js
@@ -1,0 +1,30 @@
+import {createElement} from '../../utils.js';
+import {createFilterTemplate} from './filter-tpl.js';
+
+class Filter {
+  constructor(filters) {
+    this._filters = filters;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createFilterTemplate(this._filters);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
+
+
+export {
+  Filter
+};
