+++ src/components/load-button/load-button-tpl.js
@@ -1,0 +1,14 @@
+/**
+ * Создаем разметку Кнопка Load more»
+ * @return {void}
+ */
+const createLoadButtonTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+
+export {
+  createLoadButtonTemplate
+};
