+++ src/components/site-menu/site-menu-tpl.js
@@ -1,6 +1,3 @@
-import {createElement} from '../utils.js';
-
-
 /**
  * Создаем разметку блока Меню
  * @return {void}
@@ -38,29 +35,7 @@ const createSiteMenuTemplate = () => {
   );
 };
 
-class SiteMenu {
-  constructor() {
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createSiteMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
-
 
 export {
-  SiteMenu
+  createSiteMenuTemplate
 };
