+++ src/components/site-menu/site-menu.js
@@ -1,0 +1,30 @@
+import {createElement} from '../../utils.js';
+import {createSiteMenuTemplate} from './site-menu-tpl.js';
+
+
+class SiteMenu {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createSiteMenuTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
+
+
+export {
+  SiteMenu
+};
