+++ src/components/sort/sort-tpl.js
@@ -1,6 +1,3 @@
-import {createElement} from '../utils.js';
-
-
 /**
  * Создаем разметку блока Сортировка
  * @return {string} Разметка
@@ -17,27 +14,7 @@ const createSortTemplate = () => {
   );
 };
 
-class Sort {
-  constructor() {
-    this._element = null;
-  }
 
-  getTemplate() {
-    return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
-
-
-export {Sort};
+export {
+  createSortTemplate
+};
