+++ src/components/task/task-board-tpl.js
@@ -1,0 +1,12 @@
+/**
+ * Создаем разметку блока Задач
+ * @return {string} Разметка
+ */
+const createTaskTemplate = () => {
+  return (
+    `<div class="board__tasks"></div>`
+  );
+};
+
+
+export {createTaskTemplate};
