+++ src/components/task/task-tpl.js
@@ -1,10 +1,9 @@
 import {
   MonthNames
-} from '../const.js';
+} from '../../consts.js';
 import {
-  createElement,
   formatTime
-} from '../utils.js';
+} from '../../utils.js';
 
 
 /**
@@ -73,48 +72,6 @@ const createTaskTemplate = (task) => {
   );
 };
 
-/**
- * Класс, представляющий задачу
- */
-class Task {
-  /**
-   * Создаем новую задачу
-   * @param {object} task - Задача
-   */
-  constructor(task) {
-    this._task = task;
-    this._element = null;
-  }
-
-  /**
-   * Создаём вёрстку задачи
-   * @return {string} Вёрстка задачи
-   */
-  getTemplate() {
-    return createTaskTemplate(this._task);
-  }
-
-  /**
-   * Возвращаем вёрстку задачи
-   * @return {string} Вёрстка задачи
-   */
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  /**
-   * Удаляем вёрстку задачи
-   */
-  removeElement() {
-    this._element = null;
-  }
-}
-
-
 export {
-  Task
+  createTaskTemplate
 };
